+++ scripts/portable-fixed-webview2.mjs
@@ -65,7 +65,7 @@ async function resolvePortable() {
   const packageJson = require('../package.json')
   const { version } = packageJson
 
-  const zipFile = `Clash.Verge.for.v2free_${version}_${arch}_fixed_webview2_portable.zip`
+  const zipFile = `Clash.Verge_${version}_${arch}_fixed_webview2_portable.zip`
   zip.writeZip(zipFile)
 
   console.log('[INFO]: create portable zip successfully')